+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.127] - 2025-07-31
+### Added
+* `KokoMath.Mod` for floats
+* `Randomizer` to better manipulate calls to `UnityEngine.Random`
+
 ## [0.0.126] - 2025-07-30
 ### Added
 * `FindFirstAssetByType` overloaded method to `EditorOnlyUtils`